rename project name
