+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "kit_portal_crawler/shibboleth_login"]
-	path = kit_portal_crawler/shibboleth_login
-	url = https://github.com/pddg/shibboleth_login
+[submodule "student_portal_crawler/shibboleth_login"]
+	path = student_portal_crawler/shibboleth_login
+	url = https://github.com/StudioAquatan/shibboleth_login
